Remove FIXME in TypeExpansionAnalysis: It's cache needed to be keyed by the type expansion context
